put the new branch command into the existing checkout command, but as a flag
