+++ app/cmd/checkout.go
@@ -14,31 +14,52 @@ import (
 
 const exit = "Exit"
 
+var newBranchFlag bool
+
 var checkoutCmd = &cobra.Command{
 	Use:   "cko",
 	Short: "Checkout a branch",
 	Run: func(cmd *cobra.Command, args []string) {
 
-		branchCommander := run.Commander{
-			Command: "git",
-			Args:    []string{"branch", "-a"},
+		var branchName string
+
+		if newBranchFlag {
+			question := getBranchNameQuestion()
+			err := survey.Ask(question, &branchName)
+
+			if err != nil {
+				fmt.Printf("error: %v", errors.WithMessage(err, ""))
+				return
+			}
+
+			if branchName == "exit" {
+				os.Exit(1)
+			}
+
+		} else {
+
+			branchCommander := run.Commander{
+				Command: "git",
+				Args:    []string{"branch", "-a"},
+			}
+
+			branches, err := git.GetBranches(branchCommander)
+
+			if err != nil {
+				fmt.Printf("error: %v", errors.WithMessage(err, ""))
+				return
+			}
+
+			branchName = askUser(branches)
 		}
 
-		branches, err := git.GetBranches(branchCommander)
-
-		if err != nil {
-			fmt.Printf("error: %v", errors.WithMessage(err, ""))
-			return
-		}
-
-		branchName := askUser(branches)
-
-		checkout(branchName, false)
+		checkout(branchName, newBranchFlag)
 	},
 }
 
 func init() {
 	rootCmd.AddCommand(checkoutCmd)
+	checkoutCmd.Flags().BoolVarP(&newBranchFlag, "new", "n", false, "create new branch")
 }
 
 func askUser(branches []string) string {
@@ -103,4 +124,30 @@ func checkout(branchSelection string, new bool) {
 	}
 
 	fmt.Println(result)
+}
+
+func getBranchNameQuestion() []*survey.Question {
+	var branchNameMessage = []*survey.Question{
+		{
+			Name: "branch name",
+			Prompt: &survey.Input{
+				Message: `Enter a branch name or "exit" to cancel`,
+			},
+			Validate: func(val interface{}) error {
+
+				input := strings.Replace(val.(string), " ", "", -1)
+
+				if str, ok := val.(string); !ok || len(str) != len(input) {
+					return errors.New("Branch name cannot contain spaces")
+				}
+				if str, ok := val.(string); !ok || len(str) == 0 {
+					return errors.New("Branch name can't be empty")
+				}
+
+				return nil
+			},
+		},
+	}
+
+	return branchNameMessage
 }
